tools/metrics/evaluator.py
import sys
import os

try:
    import fiftyone as fo
except ImportError:
    fo = None

class Evaluator:
    def __init__(self, dataset_name="visionstream_run", dataset_dir=None):
        self.dataset_name = dataset_name
        self.dataset_dir = dataset_dir
        
        # We will hold predictions here to compute mAP later or visualize
        self.predictions = []
        
        if fo is not None:
            # Initialize or load a FiftyOne dataset
            if fo.dataset_exists(self.dataset_name):
                self.dataset = fo.load_dataset(self.dataset_name)
                print(f"[Evaluator] Loaded existing FiftyOne dataset: {self.dataset_name}")
            else:
                self.dataset = fo.Dataset(self.dataset_name)
                print(f"[Evaluator] Created FiftyOne dataset: {self.dataset_name}")
        else:
            self.dataset = None
            print("[Evaluator] FiftyOne not installed. Visualizations and advanced mAP disabled.")

    def add_batch_results(self, images, bboxes_true, labels_true, yolo_results):
        """
        Record a batch of results.
        images: Original or preprocessed images.
        bboxes_true, labels_true: ground truth from DALI.
        yolo_results: The output from YoloInferenceNode.
        """
        # In a real environment, we'd iterate over the batch and map YOLO outputs
        # to COCO format, then optionally add them to the FiftyOne dataset.
        
        # For Phase 2 baseline, we just collect them in memory
        self.predictions.append(yolo_results)
        
    def compute_map(self):
        """
        Compute standard mean Average Precision (mAP).
        """
        if not self.predictions:
            print("[Evaluator] No predictions to evaluate.")
            return 0.0
            
        print("[Evaluator] Computing mAP...")
        # Placeholder for actual metric computation logic
        # Either using pycocotools or FiftyOne's built in evaluate_detections
        
        dummy_map = 0.42 # Simulating a 42% mAP
        return dummy_map

    def launch_dashboard(self):
        """
        Launch the FiftyOne UI.
        """
        if self.dataset is not None:
            print(f"[Evaluator] Launching FiftyOne App for dataset {self.dataset_name}...")
            session = fo.launch_app(self.dataset)
            session.wait()
        else:
            print("[Evaluator] FiftyOne not available.")
